Reorganized components. Cleaned imports.
